refactor: `narwhals.stable.v1._dtypes` -> `narwhals.dtypes._classes_v1`
Partially resolves https://github.com/narwhals-dev/narwhals/pull/3396#discussion_r2678724894
