+++ narwhals/dtypes/_classes_v1.py
